Fix google (#57)
* add message

* Add OAuth2 callback URLs and fetch user data

* refactor: 为desktop和template应用把referralCode的query，种到cookie里面 (#56)

Co-authored-by: 林峰 <lf02059301@alibaba-inc.com>

* feat: update devbox network

* Refactor OAuth and user authentication handling to include config

* Add config column to UserInfo table

This commit introduces a new JSONB column named "config" to the UserInfo
table via a database migration. The column is added to store additional
configuration settings for each user. Additionally, related code changes
are made to the application to accommodate the new column, including
updates to the Callback and CallbackVerify pages as well as backend
authentication services.

---------

Co-authored-by: taozhi <taozhi8833998@gmail.com>
Co-authored-by: 林峰 <lf02059301@alibaba-inc.com>
Co-authored-by: zjy <3161362058@qq.com>